refac: ロジック所属の整理
