+++ src/pyhees/section3_1_e.py
@@ -10,11 +10,9 @@ from functools import lru_cache
 import jjjexperiment.constants as jjj_consts
 from jjjexperiment.common import *
 from jjjexperiment.logger import log_res
-# NOTE: pyhees->jjj への依存は可能な限り絞る
-from jjjexperiment.common import get_current_injector
+# NOTE: pyhees->jjj 方向の依存はファイルスコープとしては可能な限り控える
+# インプットデータクラスとロジックを分離しているためデータクラスのみインポート可能
 from jjjexperiment.inputs.options import 床下空調ロジック
-from jjjexperiment.underfloor_ac.section3_1_e import get_Theta_uf_d_t_runup_new_ufac
-
 from jjjexperiment.underfloor_ac.inputs.common import UnderfloorAc, UfVarsDataFrame
 
 # ============================================================================
@@ -443,6 +441,7 @@ def calc_Theta(region, A_A, A_MR, A_OR, Q, r_A_ufvnt, underfloor_insulation, The
 
     if new_ufac is None:
       # 引数で渡されていないときスレッドにセットされていないかチェックする
+      from jjjexperiment.common import get_current_injector
       thread_injector = get_current_injector()
       if thread_injector is not None:
         new_ufac = thread_injector.get(UnderfloorAc)
@@ -450,8 +449,9 @@ def calc_Theta(region, A_A, A_MR, A_OR, Q, r_A_ufvnt, underfloor_insulation, The
 
     # 助走計算用床下温度
     if new_ufac is not None and new_ufac.new_ufac_flg == 床下空調ロジック.変更する:
+        from jjjexperiment.underfloor_ac.section3_1_e import get_Theta_uf_d_t_runup as jjj_ufac_get_Theta_uf_d_t_runup
         #260112 IGUCHI 新床下空調用固定値
-        Theta_uf_runup = get_Theta_uf_d_t_runup_new_ufac()
+        Theta_uf_runup = jjj_ufac_get_Theta_uf_d_t_runup()
     else:
         Theta_uf_runup = get_Theta_uf_d_t_runup(underfloor_insulation, Theta_ex_d_t)
 
@@ -700,7 +700,7 @@ def get_phi_1_A_m(m):
     """
     return get_table_e_7()[m - 1][1]
 
-
+@jjj_cloned  # 床下空調新ロジック
 def get_Theta_uf_d_t_runup(underfloor_insulation, Theta_ex_d_t):
     """<<助走計算>> 当該住戸の床下温度 (℃) (13)
