+++ src/tests/underfloor_ac/test_4_2.py
@@ -4,9 +4,8 @@ import numpy as np
 # JJJ
 from jjjexperiment.inputs.di_container import create_injector_from_json
 from jjjexperiment.inputs.common import HouseInfo
-from jjjexperiment.inputs.house_service import get_r_A_NR_uf_1F_excl_bath
-from jjjexperiment.underfloor_ac.section3_1_e import get_Theta_uf_d_t_runup_new_ufac
-from jjjexperiment.underfloor_ac.section4_2 import get_A_s_ufac_i
+from jjjexperiment.underfloor_ac.section3_1_e import get_Theta_uf_d_t_runup
+from jjjexperiment.underfloor_ac.section4_2 import get_A_s_ufac_i, get_r_A_NR_uf_1F_excl_bath
 from test_utils.utils import load_input_yaml
 
 class Test_床下空調時_共通:
@@ -41,14 +40,13 @@ class Test_床下空調時_共通:
         expected = (3.31 + 1.66 + 10.76) / 38.93  # 0.404
         assert r == pytest.approx(expected, rel=1e-4)
 
-    # TODO: 最適なテストクラスへ移動
     def test_新床下空調助走計算用床下温度(self):
         """
         get_Theta_uf_d_t_runup_new_ufac が 8760 要素の定数配列を返し、
         季節区分ごとの温度が正しいこと
         """
         # Act
-        arr = get_Theta_uf_d_t_runup_new_ufac()
+        arr = get_Theta_uf_d_t_runup()
 
         # Assert
         assert len(arr) == 24 * 365
